Add geoblocking warning banner

+++ src/components/EventModal/EventModal.js
@@ -7,6 +7,23 @@ import { Link } from "react-router-dom";
 import { shareToTwitter } from "../../utils/common";
 
 const EVENTS = {
+  "seenPopupV3": {
+    title: (
+      <>
+        <small>Upcoming Changes</small>
+      </>
+    ),
+    description: (
+      <>
+      <span>
+        Notice to Australian users of Perpetual Swaps, Perpetual Pools, MYC Staking, and the TCR to MYC Token Migration portals.
+      </span>
+      <span>
+        Please note that from 11:59 pm AEST on 16 December 2022, Australian users will be geo-blocked from accessing these subdomains. It is recommended that Australian users close out any involvement they have with these four products before this time.
+      </span>
+      </>
+    ),
+  },
   "new-earn-page": {
     title: (
       <>
@@ -55,6 +72,7 @@ export default function EventModal({
   twitterText,
   eventKey,
   hideHeader,
+  requiresConfirmation
 }) {
   const event = EVENTS[eventKey];
 
@@ -62,11 +80,14 @@ export default function EventModal({
     const hasSeenEventModal = window.localStorage.getItem(eventKey);
     if (!hasSeenEventModal) {
       setEventModalVisible(true);
-      window.localStorage.setItem(eventKey, "true");
+      if (!requiresConfirmation) {
+        window.localStorage.setItem(eventKey, "true");
+      }
     }
   }, [eventKey, setEventModalVisible]);
 
-  const onClose = () => {
+  const onContinue = () => {
+    window.localStorage.setItem(eventKey, "true");
     setEventModalVisible(false);
   };
 
@@ -85,7 +106,6 @@ export default function EventModal({
       <Styled.EventModalContent>
         <Styled.EventModalTitle>{event.title}</Styled.EventModalTitle>
         <Styled.EventModalButtonContent className="Button-content">
-          <Styled.EventModalCloseButton onClick={onClose} />
           {event.description}
           {twitterText && (
             <button className="App-button-option App-card-option" onClick={() => shareToTwitter(twitterText)}>
@@ -93,11 +113,11 @@ export default function EventModal({
             </button>
           )}
           {event.continueLink ? (
-            <Link to={event.continueLink} onClick={onClose} className="App-button-option App-card-option">
+            <Link to={event.continueLink} onClick={onContinue} className="App-button-option App-card-option">
               {event.continueLinkText ? event.continueLinkText : `Continue`}
             </Link>
           ) : (
-            <button className="App-button-option App-card-option" onClick={onClose}>
+            <button className="App-button-option App-card-option" onClick={onContinue}>
               {event.continueLinkText ? event.continueLinkText : `Continue`}
             </button>
           )}
